chore(healthcheck): service_current_status 제거 및 스키마 정합성 수정
- service_current_status: 확정 스키마에 없는 테이블 — 관련 파일 5개 제거
- health_check_logs: http_status_code (nullable) 필드 추가
- health_check_logs: response_ms long not-null → Integer nullable 수정
- HealthCheckPollService: httpStatusCode 추적 및 저장
- HealthCheckLogResponse: responseMs long → Integer
- docs/mvp.md: #33, #36, #38 MVP 범위 밖으로 이동

+++ api/src/main/java/com/nextdv/api/healthcheck/HealthCheckLogResponse.java
@@ -13,6 +13,6 @@ public class HealthCheckLogResponse {
   private UUID id;
   private UUID platformId;
   private ServiceStatus status;
-  private long responseMs;
+  private Integer responseMs;
   private Instant createdAt;
 }
